Renamed to unknown command to better match current functionality
