+++ README.md
@@ -162,7 +162,6 @@ using the matched rule and runs it. Rules enabled by default are as follows:
 * `go_run` &ndash; appends `.go` extension when compiling/running Go programs
 * `grep_recursive` &ndash; adds `-r` when you trying to `grep` directory;
 * `gulp_not_task` &ndash; fixes misspelled gulp tasks;
-* `hadoop_dfs_missing_dash` &ndash; Add the missing dash to the command.
 * `has_exists_script` &ndash; prepends `./` when script/binary exists;
 * `heroku_no_command` &ndash; fixes wrong `heroku` commands like `heroku log`;
 * `history` &ndash; tries to replace command with most similar command from history;
@@ -192,6 +191,7 @@ using the matched rule and runs it. Rules enabled by default are as follows:
 * `tsuru_login` &ndash; runs `tsuru login` if not authenticated or session expired;
 * `tsuru_not_command` &ndash; fixes wrong tsuru commands like `tsuru shell`;
 * `tmux` &ndash; fixes `tmux` commands;
+* `unknown_command` &ndash; fixes hadoop hdfs-style "unknown command" for example adds missing '-' to the command on `hdfs dfs ls`;
 * `whois` &ndash; fixes `whois` command.
 
 Enabled by default only on specific platforms:
