+++ thefuck/rules/unknown_command.py
